[build-frameworks] change - move `XcodeBuildController` from `TuistAutomation` to `TuistSupport`
